# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Interactive AI Coach System API
  version: 1.0.0
  description: |
    通用互動式教學平台 + 有證據鏈的 AI Learning Coach。

    ## 全域不變條件（SA §0.3）

    * **INV-1** `organization_id` 由 Token scope 與資源 ownership 交叉驗證得出，
      不採信 client 傳入值。
    * **INV-2** 已發布的 course version 為 immutable。
    * **INV-3** AI Coach 不得寫入成績、註冊狀態或證書。
    * **INV-4** 完成判定為 deterministic，不依賴 LLM 可用性。
    * **INV-5** Coach 正式回答必須有通過 ACL 驗證的 citation，否則回 fallback。
    * **INV-6** 學習歷史 append-only。
    * **INV-7** Elasticsearch 的 tenant filter 由 server 注入。
    * **INV-8** 所有 write endpoint 依序通過
      AuthN → RBAC → License Capability → Ownership → Audit。

    ## 錯誤處理

    資源不存在與無權限**一律回 404**（ADR-019），避免以 ID guessing 探測
    其他組織資源是否存在。403 僅用於「在可視 scope 內但缺 permission」。
  contact:
    name: Interactive AI Coach System
  license:
    name: Proprietary
servers:
  - url: https://{host}/api
    description: 已認證 API
    variables:
      host:
        default: localhost
  - url: https://{host}
    description: 公開端點（僅證書驗證與首頁 CMS）
    variables:
      host:
        default: localhost
security:
  - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: iac_session
      description: |
        HttpOnly + Secure + SameSite=Lax。已登入狀態下的所有非 GET 請求另需
        `X-CSRF-Token` header，其值須與 `iac_csrf` cookie 相同，且等於
        HMAC(SESSION_SECRET, session id)——換 session 即失效（SD §8.2）。
        公開端點（登入、密碼重設）沒有 session 可綁定，不需要此 header。
        缺少或不符回 403 `CSRF_TOKEN_INVALID`。
...
  headers:
...
    RetryAfter:
      schema: { type: integer }
  responses:
    Unauthorized:
      description: 未認證（`UNAUTHENTICATED`）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    UnprocessableEntity:
      description: |
        驗證失敗。`COURSE_VALIDATION_FAILED` 的 `details[].issue` 帶
        `RULE_*` 子代碼；`ACTIVITY_INPUT_INVALID` 由 adapter 的
        `validateInput()` 產生。
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: '`RATE_LIMITED` 或 `AI_QUOTA_EXCEEDED`'
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
              description: 穩定的錯誤代碼；使用者可見文案由前端依此在地化
            message:
              type: string
              description: 面向開發者的英文訊息
            correlation_id: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [Identity]
      security: []
      x-audit: auth.login.succeeded
      description: |
        帳號不存在與密碼錯誤回相同訊息與相近耗時（常數時間比對，THR-S-001）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: 成功，設定 session cookie }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Identity]
      x-auth-only: true
      description: rotate session token
      responses:
        '200': { description: 已更新 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [Identity]
      security: []
      x-audit: auth.password_reset.requested
      description: 常數時間回應，不透露帳號是否存在
      responses:
        '202': { description: 若帳號存在則已寄出 }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [Identity]
      security: []
      x-audit: auth.password_reset.completed
      description: |
        token 一次性；重設後撤銷該使用者所有 session、解除登入鎖定。
        token 無效、已使用或已過期回 422 `PASSWORD_RESET_TOKEN_INVALID`。
      responses:
        '204': { description: 已重設 }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...